% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferably-package.R
\docType{package}
\name{preferably}
\alias{preferably}
\alias{preferably-package}
\title{preferably}
\description{
An accessible template for pkgdown websites with dynamic light/dark mode
support
}
\details{
This is an accessible template for 'pkgdown'. It
uses two bootstrap themes, Flatly and Darkly and utilizes the
'prefers-color-scheme' CSS variable to automatically serve either of
the two based on users' operating system setting, or allowing them to
manually toggle between them.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://preferably.amirmasoudabdol.name}
  \item Report bugs at \url{https://github.com/amirmasoudabdol/preferably/issues}
}

}
\author{
Amir Masoud Abdol (i@amirmasoudabdol.name)
}
\keyword{internal}
